+++ src/main/java/com/ccapp/ccgo/dto/LoginRequestDto.java
@@ -1,0 +1,4 @@
+package com.ccapp.ccgo.dto;
+
+public class LoginRequestDto {
+}
