+++ src/main/java/com/ccapp/ccgo/exception/CustomException.java
@@ -1,0 +1,4 @@
+package com.ccapp.ccgo.exception;
+
+public class CustomException {
+}
