+++ src/main/java/com/ccapp/ccgo/exception/GlobalExceptionHandler.java
@@ -1,0 +1,4 @@
+package com.ccapp.ccgo.exception;
+
+public class GlobalExceptionHandler {
+}
