+++ src/main/java/com/ccapp/ccgo/jwt/JwtToken.java
@@ -1,0 +1,4 @@
+package com.ccapp.ccgo.jwt;
+
+public class JwtToken {
+}
